% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnet.R
\name{bnet}
\alias{bnet}
\title{Make shiny page to build bayesian network structures}
\usage{
bnet(df)
}
\arguments{
\item{df}{The data.frame passed in. No defaults.}
}
\description{
This function takes a data.frame, and spins ups a shiny one page app
where you can create bayesian networks using various algorithms
from the bnlearn package to learn the structures
}
\examples{
bnet()
}
\keyword{bnet}

