% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnets.R
\docType{package}
\name{bnets}
\alias{bnets}
\alias{bnets-package}
\title{bnets: A package to play with Bayesian Networks}
\description{
bnets combines functionalities from bnlearn, visNetwork, and shiny to make interactive bayesian networks.
}

