% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{draw_network}
\alias{draw_network}
\title{Draw the network}
\usage{
draw_network(df = wbcd, algo)
}
\arguments{
\item{df}{Preprocessed dataframe (i.e. discretized numerics and factors). Defaults to wbcd.}

\item{algo}{Name of algorithm to be used for structure learning}
}
\description{
Uses visNetwork to draw a network
}

