% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_app}
\alias{make_app}
\title{Make App}
\usage{
make_app(df = wbcd)
}
\arguments{
\item{df}{The data.frame passed in. Defaults to wbcd.}
}
\description{
Make shiny page to create bayesian networks
}

