% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wbcd}
\alias{wbcd}
\title{Wisconsin Breast Cancer Data}
\format{A data frame with 699 rows and 11 variables:
\describe{
  \item{Sample.code.number}
  \item{Clump.Thickness}
  \item{Uniformity.of.Cell.Size}
  \item{Uniformity.of.Cell.Shape}
  \item{Marginal.Adhesion}
  \item{Single.Epithelial.Cell.Size}
  \item{Bare.Nuclei}
  \item{Bland.Chromatin}
  \item{Normal.Nucleoli}
  \item{Mitoses}
  \item{Class}
}}
\source{
\url{}
}
\usage{
wbcd
}
\description{
Wisconsin Breast Cancer Data
}
\keyword{datasets}

