% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnet.R
\name{bnet}
\alias{bnet}
\title{Make App}
\usage{
bnet(df = wbcd)
}
\arguments{
\item{df}{The data.frame passed in. No defaults.}
}
\description{
Make shiny page to create create bayesian networks
}
\examples{
bnet(df)
}

